fix: remove better-sqlite3 from main process bundle
