+++ desktop_app/src/backend/server/plugins/oauth/providers.ts
@@ -1,48 +1,20 @@
 import { OAuthServerMetadata, oauthDiscoveryService } from '@backend/services/oauth-discovery';
 import log from '@backend/utils/logger';
 
-import { OAuthProviderDefinition, OAuthProviderRegistry } from './provider-interface';
-import { googleProvider } from './providers/google';
-import { linkedinBrowserProvider } from './providers/linkedin-browser';
-import { slackProvider } from './providers/slack';
-import { slackBrowserProvider } from './providers/slack-browser';
+import { OAuthProviderDefinition } from './provider-interface';
+import { oauthProviders } from './provider-registry';
 
-/**
- * Registry of OAuth providers with extensible token handling.
- * These are PUBLIC client IDs (not secrets) - safe to hardcode.
- * The secrets are stored in the OAuth proxy server.
- */
-export const oauthProviders: OAuthProviderRegistry = {
-  google: googleProvider,
-  slack: slackProvider,
-  'slack-browser': slackBrowserProvider,
-  'linkedin-browser': linkedinBrowserProvider,
-};
-
-/**
- * Get OAuth provider definition
- */
-export function getOAuthProvider(name: string): OAuthProviderDefinition {
-  const provider = oauthProviders[name.toLowerCase()];
-  if (!provider) {
-    throw new Error(`OAuth provider '${name}' not configured`);
-  }
-  return provider;
-}
-
-/**
- * Check if a provider is configured
- */
-export function hasOAuthProvider(name: string): boolean {
-  return name.toLowerCase() in oauthProviders;
-}
-
-/**
- * Get all configured provider names
- */
-export function getOAuthProviderNames(): string[] {
-  return Object.keys(oauthProviders);
-}
+// Re-export database-free functions from provider-registry
+export {
+  oauthProviders,
+  getOAuthProvider,
+  hasOAuthProvider,
+  getOAuthProviderNames,
+  googleProvider,
+  slackProvider,
+  slackBrowserProvider,
+  linkedinBrowserProvider,
+} from './provider-registry';
 
 /**
  * Helper function to get cached discovery metadata from database
@@ -148,6 +120,3 @@ export async function getOAuthProviderWithDiscovery(name: string, serverId?: str
     };
   }
 }
-
-// Re-export individual providers for direct access if needed
-export { googleProvider, slackProvider, slackBrowserProvider, linkedinBrowserProvider };
